\name{ppp2020-package}
\alias{ppp2020-package}
\alias{ppp2020}
\docType{package}
\title{
\packageTitle{ppp2020}
}
\description{
\packageDescription{ppp2020}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ppp2020}
\packageIndices{ppp2020}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ppp2020}

Maintainer: \packageMaintainer{ppp2020}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
}
