\name{ppp2020-package}
\alias{ppp2020-package}
\alias{ppp2020}
\docType{package}
\title{
\packageTitle{ppp2020}
}
\description{
\packageDescription{ppp2020}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ppp2020}
\packageIndices{ppp2020}
aa
}
\author{
\packageAuthor{ppp2020}

Maintainer: \packageMaintainer{ppp2020}
}
\references{
yes
}
ok
\keyword{ package }

\examples{
}
